Atualizando o processo ❌

+++ back-end/src/database/migrations/20211109012716-companies.js
@@ -1,17 +1,18 @@
-'use strict';
+"use strict";
 
 module.exports = {
-  up: (queryInterface, Sequelize) => queryInterface.createTable("companies", {
-    id: {
-    type: Sequelize.INTEGER,
-    primaryKey:true,
-    autoIncrement: true,
-    allowNull:false
-    },
-    cnpj: Sequelize.INTEGER,
-    name: Sequelize.STRING(100),
-    interest_rate: Sequelize.INTEGER,
- }),
-   
-  down: (queryInterface) => queryInterface.dropTable("companies")
+  up: (queryInterface, Sequelize) =>
+    queryInterface.createTable("companies", {
+      id: {
+        type: Sequelize.INTEGER,
+        primaryKey: true,
+        autoIncrement: true,
+        allowNull: false,
+      },
+      cnpj: Sequelize.INTEGER(20),
+      name: Sequelize.STRING(100),
+      interest_rate: Sequelize.INTEGER,
+    }),
+
+  down: (queryInterface) => queryInterface.dropTable("companies"),
 };
